requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlmodel==0.0.14
chromadb==0.4.18
numpy==1.26.0
pandas==2.1.3
python-dotenv==1.0.0
httpx==0.25.2
apscheduler==3.10.4
streamlit==1.28.1
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
openai==1.35.0
requests
sqlalchemy[asyncio]
asyncpg
alembic
pgvector
psycopg[binary]
google-generativeai==0.7.2
SpeechRecognition
pyttsx3
pyaudio